Refactoring: GeoDataRiverSurvey::RemoveExtensionCommand is moved to separated files.
