+++ libs/geodata/riversurvey/riversurvey.pro
@@ -125,6 +125,7 @@ HEADERS += gd_riversurvey_global.h \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.h \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.h \
            private/geodatariversurvey_moveriverpathpointcommand.h \
+           private/geodatariversurvey_removeextensioncommand.h \
            private/geodatariversurvey_setdisplaysettingcommand.h \
            private/geodatariversurveycrosssectionwindow_impl.h \
            private/geodatariversurveyproxy_setsettingcommand.h
@@ -176,6 +177,7 @@ SOURCES += geodatarivercrosssection.cpp \
            private/geodatariversurvey_mouserotaterivercrosssectioncommand.cpp \
            private/geodatariversurvey_mouseshiftriverpathcentercommand.cpp \
            private/geodatariversurvey_moveriverpathpointcommand.cpp \
+           private/geodatariversurvey_removeextensioncommand.cpp \
            private/geodatariversurvey_setdisplaysettingcommand.cpp \
            private/geodatariversurveyproxy_setsettingcommand.cpp
 RESOURCES += riversurvey.qrc
